requirements.txt
streamlit==1.31.0
google-generativeai==0.8.0
audio-recorder-streamlit==0.0.8
SpeechRecognition==3.10.4
pydub==0.25.1
gTTS==2.4.0